+++ src/ast/irbuilderbpf.h
@@ -132,9 +132,7 @@ public:
   CallInst *CreatePerCpuPtr(Value *var, Value *cpu, const Location &loc);
   CallInst *CreateThisCpuPtr(Value *var, const Location &loc);
   CallInst *CreateGetSocketCookie(Value *var, const Location &loc);
-  Value *CreateGetStrAllocation(const std::string &name,
-                                const Location &loc,
-                                uint64_t pad = 0);
+  Value *CreateGetStrAllocation(const std::string &name, const Location &loc);
   Value *CreateGetFmtStringArgsAllocation(StructType *struct_type,
                                           const std::string &name,
                                           const Location &loc);