+++ src/globalvars.cpp
@@ -245,7 +245,7 @@ SizedType GlobalVars::get_sized_type(const std::string &global_var_name,
 
   if (global_var_name == GET_STR_BUFFER) {
     assert(resources.str_buffers > 0);
-    const auto max_strlen = bpftrace_config.max_strlen;
+    const auto max_strlen = bpftrace_config.pad_max_strlen();
     return make_rw_type(resources.str_buffers,
                         CreateArray(max_strlen, CreateInt8()));
   }